+++ src/Akeneo/Pim/Automation/DataQualityInsights/back/Infrastructure/Elasticsearch/Filter/QualityScoreFilter.php
@@ -4,8 +4,10 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Automation\DataQualityInsights\Infrastructure\Elasticsearch\Filter;
 
+use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\Rank;
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Filter\Field\AbstractFieldFilter;
 use Akeneo\Pim\Enrichment\Component\Product\Query\Filter\FieldFilterInterface;
+use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyException;
 use Akeneo\Tool\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 
 /**
@@ -16,7 +18,7 @@ class QualityScoreFilter extends AbstractFieldFilter implements FieldFilterInter
 {
     public function __construct()
     {
-        $this->supportedFields = ['data_quality_insights_score'];
+        $this->supportedFields = ['data_quality_insights_score', 'quality_score'];
         $this->supportedOperators = ['IN'];
     }
 
@@ -31,6 +33,23 @@ class QualityScoreFilter extends AbstractFieldFilter implements FieldFilterInter
 
         if (!is_array($values)) {
             throw InvalidPropertyTypeException::arrayExpected($field, static::class, $values);
+        }
+
+        if (null === $channel) {
+            throw InvalidPropertyException::dataExpected($field, 'a channel', static::class);
+        }
+
+        if (null === $locale) {
+            throw InvalidPropertyException::dataExpected($field, 'a locale', static::class);
+        }
+
+        try {
+            $values = array_map(function ($value) {
+                $rank = is_numeric($value) ? Rank::fromInt(intval($value)) : Rank::fromLetter(strval($value));
+                return $rank->toInt();
+            }, $values);
+        } catch (\InvalidArgumentException $exception) {
+            throw InvalidPropertyException::dataExpected($field, sprintf('values among "%s"', implode('", "', Rank::LETTERS_MAPPING)), static::class);
         }
 
         $clause = [